requirements.txt
Django==1.10.4
Pillow==4.0.0
djangorestframework==3.5.3
django-rest-auth==0.8.2
django-cors-headers==1.2.0
djangorestframework-camel-case==0.2.0
-e git://github.com/vechnoe/chakert.git@0.2.1#egg=chakert
drf-nested-routers==0.11.1
python-dateutil==2.6.0
django-wysiwyg-redactor==0.5.0
requests==2.12.3
psycopg2==2.6.2
pytz==2016.10
pytils==0.3
aiofiles==0.3.2
aiohttp==3.0.1
celery==4.0
pyyaml==3.12
setproctitle==1.1.9

pytest==3.0.4
flower==0.9.1
flake8==3.0.3
coverage==4.2
factory-boy==2.7.0
ipython>=6.x.x
raven==5.32.0

gunicorn==19.4.5
